Split effect handlers into multiple files
