+++ lib/gamedata/summon.txt
@@ -37,8 +37,8 @@
 # 'desc' is the description of the summon type used in messages.
 
 # Note that the KIN summon is a special case which has special code in
-# src/effects.c to handle it, and UNIQUE and WRAITH have code in
-# src/mon-summon.c that exclude them from being used for the "call" type of
+# effect_handler_SUMMON() to handle it, and UNIQUE and WRAITH have code in
+# summon_specific() that exclude them from being used for the "call" type of
 # summon (where summoned monsters are chosen from those already on the level
 # rather than newly created).
