+++ src/win/vs2019/Angband.vcxproj
@@ -178,6 +178,8 @@ xcopy $(MSBuildProjectDirectory)\lib\* $(OutDir)lib\ /DESY /exclude:$(MSBuildPro
     <ClCompile Include="src\cmd-wizard.c" />
     <ClCompile Include="src\datafile.c" />
     <ClCompile Include="src\debug.c" />
+    <ClCompile Include="src\effect-handler-attack.c" />
+    <ClCompile Include="src\effect-handler-general.c" />
     <ClCompile Include="src\effects-info.c" />
     <ClCompile Include="src\effects.c" />
     <ClCompile Include="src\game-event.c" />
